requirements.txt
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
